LayoutTests/gamepad/gamepad-api-expected.txt

PASS navigator.webkitGamepads is defined.
PASS successfullyParsed is true

TEST COMPLETE
Make sure the main polling access point exists on navigator.
